refactor(frontend): enforce one type per file across shared libs (US-000)
Extract all interfaces, types, and enums from service files into
dedicated files. Splits recipe/shopping/auth API DTOs, pagination
types, language code type, and AuthUser interface into individual
files with updated barrel re-exports.
