+++ client/libs/shared/api-client/src/index.ts
@@ -1,26 +1,26 @@
 export { AuthApiService } from './lib/auth-api.service';
-export {
-  RecipeApiService,
-  type ImportRecipeRequest,
-  type ImportRecipeResponse,
-  type ImportStreamEvent,
-  type ExtractedIngredient,
-  type ExtractedStep,
-  type SaveRecipeRequest,
-  type UpdateRecipeRequest,
-  type SavedRecipeResponse,
-  type RecipeListItem,
-  type RecipeListResponse,
-  type GetRecipesParams,
-  type RecipeDetail,
-  type RecipeIngredient,
-  type RecipeStep,
-} from './lib/recipe-api.service';
-export {
-  ShoppingApiService,
-  type CreateShoppingListItem,
-  type CreateShoppingListRequest,
-  type ShoppingListItemResponse,
-  type ShoppingListDetail,
-  type ShoppingListSummary,
-} from './lib/shopping-api.service';
+export type { RegisterRequest } from './lib/register-request';
+export type { RegisterResponse } from './lib/register-response';
+export type { ResendVerificationRequest } from './lib/resend-verification-request';
+export type { ResendVerificationResponse } from './lib/resend-verification-response';
+export { RecipeApiService } from './lib/recipe-api.service';
+export type { ImportRecipeRequest } from './lib/import-recipe-request';
+export type { ImportRecipeResponse } from './lib/import-recipe-response';
+export type { ImportStreamEvent } from './lib/import-stream-event';
+export type { ExtractedIngredient } from './lib/extracted-ingredient';
+export type { ExtractedStep } from './lib/extracted-step';
+export type { SaveRecipeRequest } from './lib/save-recipe-request';
+export type { UpdateRecipeRequest } from './lib/update-recipe-request';
+export type { SavedRecipeResponse } from './lib/saved-recipe-response';
+export type { RecipeListItem } from './lib/recipe-list-item';
+export type { RecipeListResponse } from './lib/recipe-list-response';
+export type { GetRecipesParams } from './lib/get-recipes-params';
+export type { RecipeDetail } from './lib/recipe-detail';
+export type { RecipeIngredient } from './lib/recipe-ingredient';
+export type { RecipeStep } from './lib/recipe-step';
+export { ShoppingApiService } from './lib/shopping-api.service';
+export type { CreateShoppingListItem } from './lib/create-shopping-list-item';
+export type { CreateShoppingListRequest } from './lib/create-shopping-list-request';
+export type { ShoppingListItemResponse } from './lib/shopping-list-item-response';
+export type { ShoppingListDetail } from './lib/shopping-list-detail';
+export type { ShoppingListSummary } from './lib/shopping-list-summary';
