+++ client/libs/shared/api-client/src/lib/create-shopping-list-item.ts
@@ -1,0 +1,5 @@
+export interface CreateShoppingListItem {
+  name: string;
+  amount: number | null;
+  unit: string | null;
+}
