+++ client/libs/shared/api-client/src/lib/create-shopping-list-request.ts
@@ -1,0 +1,7 @@
+import type { CreateShoppingListItem } from './create-shopping-list-item';
+
+export interface CreateShoppingListRequest {
+  title: string;
+  items: CreateShoppingListItem[];
+  recipeIdentifier?: string;
+}
