+++ client/libs/shared/api-client/src/lib/extracted-ingredient.ts
@@ -1,0 +1,5 @@
+export interface ExtractedIngredient {
+  name: string;
+  amount: number | null;
+  unit: string | null;
+}
