+++ client/libs/shared/api-client/src/lib/extracted-step.ts
@@ -1,0 +1,4 @@
+export interface ExtractedStep {
+  number: number;
+  description: string;
+}
