+++ client/libs/shared/api-client/src/lib/get-recipes-params.ts
@@ -1,0 +1,6 @@
+import type { PaginationParams } from '@yumney/shared/models';
+
+export interface GetRecipesParams extends PaginationParams {
+  sortBy?: 'Name' | 'Date';
+  search?: string;
+}
