+++ client/libs/shared/api-client/src/lib/import-recipe-request.ts
@@ -1,0 +1,3 @@
+export interface ImportRecipeRequest {
+  url: string;
+}
