+++ client/libs/shared/api-client/src/lib/import-stream-event.ts
@@ -1,0 +1,4 @@
+export interface ImportStreamEvent {
+  type: 'status' | 'chunk' | 'done' | 'fail';
+  data: string;
+}
