+++ client/libs/shared/api-client/src/lib/recipe-ingredient.ts
@@ -1,0 +1,5 @@
+export interface RecipeIngredient {
+  name: string;
+  amount: number | null;
+  unit: string | null;
+}
