+++ client/libs/shared/api-client/src/lib/recipe-list-item.ts
@@ -1,0 +1,12 @@
+export interface RecipeListItem {
+  identifier: string;
+  title: string;
+  description: string | null;
+  servings: number | null;
+  prepTimeMinutes: number | null;
+  cookTimeMinutes: number | null;
+  difficulty: string | null;
+  imageUrl: string | null;
+  createdAt: string;
+  tags: string[];
+}
