+++ client/libs/shared/api-client/src/lib/recipe-list-response.ts
@@ -1,0 +1,4 @@
+import type { PagedResponse } from '@yumney/shared/models';
+import type { RecipeListItem } from './recipe-list-item';
+
+export type RecipeListResponse = PagedResponse<RecipeListItem>;
