+++ client/libs/shared/api-client/src/lib/recipe-step.ts
@@ -1,0 +1,4 @@
+export interface RecipeStep {
+  number: number;
+  description: string;
+}
