+++ client/libs/shared/api-client/src/lib/register-request.ts
@@ -1,0 +1,5 @@
+export interface RegisterRequest {
+  email: string;
+  password: string;
+  displayName: string;
+}
