+++ client/libs/shared/api-client/src/lib/register-response.ts
@@ -1,0 +1,3 @@
+export interface RegisterResponse {
+  message: string;
+}
