+++ client/libs/shared/api-client/src/lib/resend-verification-request.ts
@@ -1,0 +1,3 @@
+export interface ResendVerificationRequest {
+  email: string;
+}
