+++ client/libs/shared/api-client/src/lib/resend-verification-response.ts
@@ -1,0 +1,3 @@
+export interface ResendVerificationResponse {
+  message: string;
+}
