+++ client/libs/shared/api-client/src/lib/saved-recipe-response.ts
@@ -1,0 +1,5 @@
+export interface SavedRecipeResponse {
+  identifier: string;
+  title: string;
+  createdAt: string;
+}
