+++ client/libs/shared/api-client/src/lib/shopping-list-detail.ts
@@ -1,0 +1,9 @@
+import type { ShoppingListItemResponse } from './shopping-list-item-response';
+
+export interface ShoppingListDetail {
+  identifier: string;
+  title: string;
+  recipeIdentifier: string | null;
+  createdAt: string;
+  items: ShoppingListItemResponse[];
+}
