+++ client/libs/shared/api-client/src/lib/shopping-list-item-response.ts
@@ -1,0 +1,7 @@
+export interface ShoppingListItemResponse {
+  identifier: string;
+  name: string;
+  amount: number | null;
+  unit: string | null;
+  isChecked: boolean;
+}
