+++ client/libs/shared/api-client/src/lib/shopping-list-summary.ts
@@ -1,0 +1,6 @@
+export interface ShoppingListSummary {
+  identifier: string;
+  title: string;
+  itemCount: number;
+  createdAt: string;
+}
