+++ client/libs/shared/api-client/src/lib/update-recipe-request.ts
@@ -1,0 +1,12 @@
+export interface UpdateRecipeRequest {
+  title: string;
+  description: string | null;
+  ingredients: { name: string; amount: number | null; unit: string | null }[];
+  steps: { number: number; description: string }[];
+  servings: number | null;
+  prepTimeMinutes: number | null;
+  cookTimeMinutes: number | null;
+  difficulty: string | null;
+  imageUrl: string | null;
+  tags?: string[];
+}
