+++ client/libs/shared/auth/src/lib/auth-user.ts
@@ -1,0 +1,6 @@
+export interface AuthUser {
+  sub: string;
+  email: string;
+  preferredUsername: string;
+  roles: string[];
+}
