+++ client/libs/shared/models/src/index.ts
@@ -1,5 +1,6 @@
 export { mapHttpError, type HttpErrorMap } from './lib/http-error-utils';
-export { type PagedResponse, type PaginationParams } from './lib/pagination';
+export { type PagedResponse } from './lib/paged-response';
+export { type PaginationParams } from './lib/pagination-params';
 export { urlValidator, passwordsMatchValidator } from './lib/validators';
 export { VALIDATION } from './lib/validation-constants';
 export { scaleIngredients, type ScalableIngredient } from './lib/scale-ingredients';
@@ -7,12 +8,8 @@ export { UI } from './lib/ui-constants';
 export { createAsyncState, type AsyncState } from './lib/async-state';
 export { TranslocoHttpLoader } from './lib/transloco-loader';
 export { IS_STANDALONE } from './lib/federation-context';
-export {
-  LanguageService,
-  type LanguageCode,
-  SUPPORTED_LANGUAGES,
-  DEFAULT_LANGUAGE,
-} from './lib/language.service';
+export { LanguageService } from './lib/language.service';
+export { type LanguageCode, SUPPORTED_LANGUAGES, DEFAULT_LANGUAGE } from './lib/language-code';
 export {
   mapToSaveRecipeRequest,
   mapToUpdateRecipeRequest,
