+++ client/libs/shared/models/src/lib/language-code.ts
@@ -1,0 +1,4 @@
+export type LanguageCode = 'en' | 'de';
+
+export const SUPPORTED_LANGUAGES: readonly LanguageCode[] = ['en', 'de'] as const;
+export const DEFAULT_LANGUAGE: LanguageCode = 'en';
