+++ client/libs/shared/models/src/lib/paged-response.ts
@@ -1,0 +1,6 @@
+export interface PagedResponse<T> {
+  items: T[];
+  totalCount: number;
+  page: number;
+  pageSize: number;
+}
