+++ client/libs/shared/models/src/lib/pagination-params.ts
@@ -1,10 +1,3 @@
-export interface PagedResponse<T> {
-  items: T[];
-  totalCount: number;
-  page: number;
-  pageSize: number;
-}
-
 export interface PaginationParams {
   page?: number;
   pageSize?: number;
